+++ android/src/main/java/com/reactnativestripesdk/PaymentMethodCreateParamsFactory.kt
@@ -32,6 +32,7 @@ class PaymentMethodCreateParamsFactory(
         PaymentMethod.Type.AuBecsDebit -> createAuBecsDebitPaymentConfirmParams()
         PaymentMethod.Type.Klarna -> createKlarnaPaymentConfirmParams()
         PaymentMethod.Type.USBankAccount -> createUSBankAccountPaymentConfirmParams()
+        PaymentMethod.Type.PayPal -> createPayPalPaymentConfirmParams()
         else -> {
           throw Exception("This paymentMethodType is not supported yet")
         }
@@ -52,6 +53,7 @@ class PaymentMethodCreateParamsFactory(
         PaymentMethod.Type.SepaDebit -> createSepaPaymentSetupParams()
         PaymentMethod.Type.AuBecsDebit -> createAuBecsDebitPaymentSetupParams()
         PaymentMethod.Type.USBankAccount -> createUSBankAccountPaymentSetupParams()
+        PaymentMethod.Type.PayPal -> createPayPalPaymentSetupParams()
         else -> {
           throw Exception("This paymentMethodType is not supported yet")
         }
@@ -467,6 +469,10 @@ class PaymentMethodCreateParamsFactory(
     }
   }
 
+  @Throws(PaymentMethodCreateParamsException::class)
+  private fun createPayPalPaymentSetupParams(): ConfirmSetupIntentParams {
+    throw PaymentMethodCreateParamsException("PayPal is not yet supported through SetupIntents.")
+  }
 
   @Throws(PaymentMethodCreateParamsException::class)
   private fun createKlarnaPaymentConfirmParams(): ConfirmPaymentIntentParams {
@@ -507,6 +513,14 @@ class PaymentMethodCreateParamsFactory(
         paymentMethodType = PaymentMethod.Type.USBankAccount
       )
     }
+  }
+
+  @Throws(PaymentMethodCreateParamsException::class)
+  private fun createPayPalPaymentConfirmParams(): ConfirmPaymentIntentParams {
+    return ConfirmPaymentIntentParams.createWithPaymentMethodCreateParams(
+      paymentMethodCreateParams = PaymentMethodCreateParams.createPayPal(null),
+      clientSecret = clientSecret,
+    )
   }
 
   @Throws(PaymentMethodCreateParamsException::class)